+++ cnd/src/libp2p_bam/handler.rs
@@ -245,11 +245,15 @@ fn poll_substreams<S: Display + Advance>(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::libp2p_bam::test_harness::{
-        request_with_no_headers, setup_substream, setup_substream_with_json_codec, IntoEventStream,
-        IntoFutureWithResponse, WaitForFrame,
+    use crate::{
+        bam_ext::ToBamHeader,
+        libp2p_bam::test_harness::{
+            request_with_no_headers, setup_substream, setup_substream_with_json_codec,
+            IntoEventStream, IntoFutureWithResponse, WaitForFrame,
+        },
+        swap_protocols::rfc003::messages::Decision,
     };
-    use bam::{frame::Header, Status};
+    use bam::frame::Header;
     use futures::{Future, Sink, Stream};
     use libp2p::swarm::ProtocolsHandlerEvent;
     use spectral::prelude::*;
@@ -269,16 +273,29 @@ mod tests {
         let dialer = runtime.block_on(send).unwrap();
 
         // and we provide an answer
-        let future = handler.into_future_with_response(Response::new(Status::OK(127)));
+        let future = handler.into_future_with_response(
+            Response::empty().with_header(
+                "decision",
+                Decision::Declined
+                    .to_bam_header()
+                    .expect("Decision should not fail to serialize"),
+            ),
+        );
         runtime.spawn(future);
 
         // then we send the response back to the dialer
         let response = runtime.block_on(dialer.wait_for_frame());
 
-        assert_that(&response)
-            .is_ok()
-            .is_some()
-            .is_equal_to(Response::new(Status::OK(127)).into_frame());
+        assert_that(&response).is_ok().is_some().is_equal_to(
+            Response::empty()
+                .with_header(
+                    "decision",
+                    Decision::Declined
+                        .to_bam_header()
+                        .expect("Decision should not fail to serialize"),
+                )
+                .into_frame(),
+        );
     }
 
     #[test]
@@ -586,45 +603,6 @@ mod tests {
         matches::assert_matches!(
             events.get(0),
             Some(ProtocolsHandlerEvent::Custom(ProtocolOutEvent::Error(Error::UnexpectedFrame(_))))
-        )
-    }
-
-    #[test]
-    fn given_an_outbound_request_when_malformed_response_should_emit_malformed_frame() {
-        let mut runtime = tokio::runtime::Runtime::new().unwrap();
-        let (dialer, listener) = runtime
-            .block_on(setup_substream(
-                JsonFrameCodec::default(),
-                LinesCodec::new(),
-            ))
-            .unwrap();
-        let mut handler = BamHandler::new(request_with_no_headers("PING"));
-
-        // given an outbound substream
-        let (sender, _receiver) = oneshot::channel();
-        handler.inject_fully_negotiated_outbound(
-            dialer,
-            ProtocolOutboundOpenInfo::Message(OutboundMessage::Request(PendingOutboundRequest {
-                request: OutboundRequest::new("PING"),
-                channel: sender,
-            })),
-        );
-
-        // when receiving a known type (REQUEST) that is unexpected
-        let send = listener
-            .send(r#"{"type": "RESPONSE", "payload":{}}"#.to_owned())
-            .map(|_| ())
-            .map_err(|_| ());
-        let _ = runtime.spawn(send);
-
-        let events = runtime
-            .block_on(handler.into_event_stream().take(1).collect())
-            .unwrap();
-
-        // then
-        matches::assert_matches!(
-            events.get(0),
-            Some(ProtocolsHandlerEvent::Custom(ProtocolOutEvent::Error(Error::MalformedFrame(_))))
         )
     }