+++ internal/bam/src/api.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, PartialEq, Copy, Clone)]
-pub enum Status {
-    OK(u8),
-    SE(u8),
-    RE(u8),
-}