+++ internal/bam/src/frame/macros.rs
@@ -18,8 +18,20 @@ macro_rules! try_header {
 macro_rules! header {
     ($e:expr) => {
         header_internal!($e, {
-            log::info!("Header was not present, early returning with error response (SE00)!");
-            return Box::new(futures::future::ok(Response::new(Status::SE(0))));
+            log::info!("Header was not present, early returning with decline response!");
+            let decline_body = DeclineResponseBody {
+                reason: Some(SwapDeclineReason::MissingMandatoryHeader),
+            };
+
+            return Box::new(futures::future::ok(Response::empty().with_header(
+                "decision",
+                Decision::Declined
+                    .to_bam_header()
+                    .expect("Decision should not fail to serialize"),
+            )
+            .with_body(serde_json::to_value(decline_body).expect(
+                "decline body should always serialize into serde_json::Value",
+            ))));
         })
     };
 }
@@ -30,8 +42,20 @@ macro_rules! body {
         match $e {
             Ok(body) => body,
             Err(e) => {
-                log::error!("Failed to deserialize body: {:?}", e);
-                return Box::new(futures::future::ok(Response::new(Status::SE(0))));
+                log::error!("Failed to deserialize body because of unexpected field: {:?}", e);
+                let decline_body = DeclineResponseBody {
+                    reason: Some(SwapDeclineReason::BadJsonField),
+                };
+
+                return Box::new(futures::future::ok(Response::empty().with_header(
+                    "decision",
+                    Decision::Declined
+                        .to_bam_header()
+                        .expect("Decision should not fail to serialize"),
+                )
+                .with_body(serde_json::to_value(decline_body).expect(
+                    "decline body should always serialize into serde_json::Value",
+                ))));
             }
         }
     };
@@ -43,8 +67,22 @@ macro_rules! header_internal {
         match $e {
             Some(Ok(header)) => header,
             Some(Err(e)) => {
-                log::error!("Failed to deserialize header: {:?}", e);
-                return Box::new(futures::future::ok(Response::new(Status::SE(0))));
+                log::error!("Failed to deserialize header because of unexpected field: {:?}", e);
+
+                let decline_body = DeclineResponseBody {
+                    reason: Some(SwapDeclineReason::BadJsonField),
+                };
+
+                return Box::new(futures::future::ok(Response::empty().with_header(
+                    "decision",
+                    Decision::Declined
+                        .to_bam_header()
+                        .expect("Decision should not fail to serialize"),
+                )
+                .with_body(serde_json::to_value(decline_body).expect(
+                    "decline body should always serialize into serde_json::Value",
+                ))));
+
             },
             None => $none,
         }